+++ src/pages/footer/footer.component.tsx
@@ -13,27 +13,27 @@ const socialLinks = [
 	{
 		name: 'Instagram',
 		icon: <FaInstagram />,
-		link: 'http://localhost:3000/',
+		link: 'https://www.instagram.com/',
 	},
 	{
 		name: 'Twitter',
 		icon: <FaTwitter />,
-		link: 'http://localhost:3000/',
+		link: 'https://twitter.com/',
 	},
 	{
 		name: 'Facebook',
 		icon: <FaFacebookF />,
-		link: 'http://localhost:3000/',
+		link: 'https://www.facebook.com/',
 	},
 	{
 		name: 'Discord',
 		icon: <FaDiscord />,
-		link: 'http://localhost:3000/',
+		link: 'https://discord.com/',
 	},
 	{
 		name: 'YouTube',
 		icon: <FaYoutube />,
-		link: 'http://localhost:3000/',
+		link: 'https://www.youtube.com/',
 	},
 ];
 
@@ -45,6 +45,8 @@ const Footer: FC = () => {
 					<a
 						key={link.name}
 						href={link.link}
+						target='_blank'
+						rel='noopener noreferrer'
 						className='social-icon-link'
 						aria-label={link.name}
 					>