+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # WIP
 
+# v1.1.0 - 20th October 2023
+
+- Updated social (icon) links
+- Updated FAG index tooltip styles (mobile / tablet)
+- Updated 'Market' section to store row and page number state when returning from other page
+- Enhanced user navigation with scroll position retention for coin list
+- Introduced CoinGecko context and refactored Home & Market components to reduce number of API calls made
+
 # v1.0.0 - 1st October 2023
 
 - Initial project setup using create-react-app with TypeScript and SASS